Add very basic plugin structure for creating new printers
This moves the OctoPrint printer implementation to being a default
plugin, and allows other printers types to be added as new printer
plugins. Currently, only one printer plugin should be active at a time,
and you must restart OctoPrint after changing the active printer plugin.

AE-626
