+++ src/octoprint/plugin/__init__.py
@@ -103,7 +103,9 @@ def plugin_manager(init=False, plugin_folders=None, plugin_types=None, plugin_en
 				                EventHandlerPlugin,
 				                SlicerPlugin,
 				                AppPlugin,
-				                ProgressPlugin]
+				                ProgressPlugin,
+				                PrinterPlugin,
+				                MachineComPrintCallbackPlugin]
 			if plugin_entry_points is None:
 				plugin_entry_points = "octoprint.plugin"
 			if plugin_disabled_list is None:
