+++ src/octoprint/plugin/types.py
@@ -1121,3 +1121,343 @@ class AppPlugin(OctoPrintPlugin):
 	def get_additional_apps(self):
 		return []
 
+class PrinterPlugin(OctoPrintPlugin):
+	"""
+	The `PrinterPlugin` mixin allows plugin developers to specify their own printer controllers for
+	adding support for potentially any printer-like device to OctoPrint.
+	Without one (and only one) printer plugin enabled, OctoPrint will not function properly.
+
+	The OctoPrintPrinter is the default PrinterPlugin.
+	"""
+
+	def startup(self, fileManager, analysisQueue, printerProfileManager):
+		"""
+		Initilizes the plugin for use.
+		"""
+		pass
+
+	def connect(self, port=None, baudrate=None, profile=None):
+		"""
+		Connects to the printer, using the specified serial ``port``, ``baudrate`` and printer ``profile``. If a
+		connection is already established, that connection will be closed prior to connecting anew with the provided
+		parameters.
+
+		Arguments:
+		    port (str): Name of the serial port to connect to. If not provided, an auto detection will be attempted.
+		    baudrate (int): Baudrate to connect with. If not provided, an auto detection will be attempted.
+		    profile (str): Name of the printer profile to use for this connection. If not provided, the default
+			will be retrieved from the :class:`PrinterProfileManager`.
+		"""
+		pass
+
+	def disconnect(self):
+		"""
+		Disconnects from the printer. Does nothing if no connection is currently established.
+		"""
+		raise NotImplementedError()
+
+	def get_transport(self):
+		"""
+		Returns the communication layer's transport object, if a connection is currently established.
+
+		Note that this doesn't have to necessarily be a :class:`serial.Serial` instance, it might also be something
+		different, so take care to do instance checks before attempting to access any properties or methods.
+
+		Returns:
+		    object: The communication layer's transport object
+		"""
+		raise NotImplementedError()
+
+	def fake_ack(self):
+		"""
+		Fakes an acknowledgement for the communication layer. If the communication between OctoPrint and the printer
+		gets stuck due to lost "ok" responses from the server due to communication issues, this can be used to get
+		things going again.
+		"""
+		raise NotImplementedError()
+
+	def commands(self, commands):
+		"""
+		Sends the provided ``commands`` to the printer.
+
+		Arguments:
+		    commands (str, list): The commands to send. Might be a single command provided just as a string or a list
+			of multiple commands to send in order.
+		"""
+		raise NotImplementedError()
+
+	def script(self, name, context=None):
+		"""
+		Sends the GCODE script ``name`` to the printer.
+
+		The script will be run through the template engine, the rendering context can be extended by providing a
+		``context`` with additional template variables to use.
+
+		If the script is unknown, an :class:`UnknownScriptException` will be raised.
+
+		Arguments:
+		    name (str): The name of the GCODE script to render.
+		    context (dict): An optional context of additional template variables to provide to the renderer.
+
+		Raises:
+		    UnknownScriptException: There is no GCODE script with name ``name``
+		"""
+		raise NotImplementedError()
+
+	def jog(self, axis, amount):
+		"""
+		Jogs the specified printer ``axis`` by the specified ``amount`` in mm.
+
+		Arguments:
+		    axis (str): The axis to jog, will be converted to lower case, one of "x", "y", "z" or "e"
+		    amount (int, float): The amount by which to jog in mm
+		"""
+		raise NotImplementedError()
+
+	def home(self, axes):
+		"""
+		Homes the specified printer ``axes``.
+
+		Arguments:
+		    axes (str, list): The axis or axes to home, each of which must converted to lower case must match one of
+			"x", "y", "z" and "e"
+		"""
+		raise NotImplementedError()
+
+	def extrude(self, amount):
+		"""
+		Extrude ``amount`` milimeters of material from the tool.
+
+		Arguments:
+		    amount (int, float): The amount of material to extrude in mm
+		"""
+		raise NotImplementedError()
+
+	def change_tool(self, tool):
+		"""
+		Switch the currently active ``tool`` (for which extrude commands will apply).
+
+		Arguments:
+		    tool (str): The tool to switch to, matching the regex "tool[0-9]+" (e.g. "tool0", "tool1", ...)
+		"""
+		raise NotImplementedError()
+
+	def set_temperature(self, heater, value):
+		"""
+		Sets the target temperature on the specified ``heater`` to the given ``value`` in celsius.
+
+		Arguments:
+		    heater (str): The heater for which to set the target temperature. Either "bed" for setting the bed
+			temperature or something matching the regular expression "tool[0-9]+" (e.g. "tool0", "tool1", ...) for
+			the hotends of the printer
+		    value (int, float): The temperature in celsius to set the target temperature to.
+		"""
+		raise NotImplementedError()
+
+	def set_temperature_offset(self, offsets=None):
+		"""
+		Sets the temperature ``offsets`` to apply to target temperatures red from a GCODE file while printing.
+
+		Arguments:
+		    offsets (dict): A dictionary specifying the offsets to apply. Keys must match the format for the ``heater``
+			parameter to :func:`set_temperature`, so "bed" for the offset for the bed target temperature and
+			"tool[0-9]+" for the offsets to the hotend target temperatures.
+		"""
+		raise NotImplementedError()
+
+	def feed_rate(self, factor):
+		"""
+		Sets the ``factor`` for the printer's feed rate.
+
+		Arguments:
+		    factor (int, float): The factor for the feed rate to send to the firmware. Percentage expressed as either an
+		    int between 0 and 100 or a float between 0 and 1.
+		"""
+		raise NotImplementedError()
+
+	def flow_rate(self, factor):
+		"""
+		Sets the ``factor`` for the printer's flow rate.
+
+		Arguments:
+		    factor (int, float): The factor for the flow rate to send to the firmware. Percentage expressed as either an
+		    int between 0 and 100 or a float between 0 and 1.
+		"""
+		raise NotImplementedError()
+
+	def select_file(self, path, sd, printAfterSelect=False):
+		"""
+		Selects the specified ``path`` for printing, specifying if the file is to be found on the ``sd`` or not.
+		Optionally can also directly start the print after selecting the file.
+
+		Arguments:
+		    path (str): The path to select for printing. Either an absolute path (local file) or a
+		"""
+		raise NotImplementedError()
+
+	def unselect_file(self):
+		"""
+		Unselects and currently selected file.
+		"""
+		raise NotImplementedError()
+
+	def start_print(self):
+		"""
+		Starts printing the currently selected file. If no file is currently selected, does nothing.
+		"""
+		raise NotImplementedError()
+
+	def toggle_pause_print(self):
+		"""
+		Pauses the current print job if it is currently running or resumes it if it is currently paused.
+		"""
+		raise NotImplementedError()
+
+	def cancel_print(self):
+		"""
+		Cancels the current print job.
+		"""
+		raise NotImplementedError()
+
+	def get_state_string(self):
+		"""
+		Returns:
+		     (str) A human readable string corresponding to the current communication state.
+		"""
+		raise NotImplementedError()
+
+	def get_current_data(self):
+		"""
+		Returns:
+		    (dict) The current state data.
+		"""
+		raise NotImplementedError()
+
+	def get_current_job(self):
+		"""
+		Returns:
+		    (dict) The data of the current job.
+		"""
+		raise NotImplementedError()
+
+	def get_current_temperatures(self):
+		"""
+		Returns:
+		    (dict) The current temperatures.
+		"""
+		raise NotImplementedError()
+
+	def get_temperature_history(self):
+		"""
+		Returns:
+		    (list) The temperature history.
+		"""
+		raise NotImplementedError()
+
+	def get_current_connection(self):
+		"""
+		Returns:
+		    (tuple) The current connection information as a 4-tuple ``(connection_string, port, baudrate, printer_profile)``.
+			If the printer is currently not connected, the tuple will be ``("Closed", None, None, None)``.
+		"""
+		raise NotImplementedError()
+
+	def is_closed_or_error(self):
+		"""
+		Returns:
+		    (boolean) Whether the printer is currently disconnected and/or in an error state.
+		"""
+		raise NotImplementedError()
+
+	def is_operational(self):
+		"""
+		Returns:
+		    (boolean) Whether the printer is currently connected and available.
+		"""
+		raise NotImplementedError()
+
+	def is_printing(self):
+		"""
+		Returns:
+		    (boolean) Whether the printer is currently printing.
+		"""
+		raise NotImplementedError()
+
+	def is_paused(self):
+		"""
+		Returns:
+		    (boolean) Whether the printer is currently paused.
+		"""
+		raise NotImplementedError()
+
+	def is_error(self):
+		"""
+		Returns:
+		    (boolean) Whether the printer is currently in an error state.
+		"""
+		raise NotImplementedError()
+
+	def is_ready(self):
+		"""
+		Returns:
+		    (boolean) Whether the printer is currently operational and ready for new print jobs (not printing).
+		"""
+		raise NotImplementedError()
+
+	def register_callback(self, callback):
+		"""
+		Registers a :class:`PrinterCallback` with the instance.
+
+		Arguments:
+		    callback (PrinterCallback): The callback object to register.
+		"""
+		raise NotImplementedError()
+
+	def unregister_callback(self, callback):
+		"""
+		Unregisters a :class:`PrinterCallback` from the instance.
+
+		Arguments:
+		    callback (PrinterCallback): The callback object to unregister.
+		"""
+		raise NotImplementedError()
+
+class MachineComPrintCallbackPlugin(OctoPrintPlugin):
+	def on_comm_log(self, message):
+		pass
+
+	def on_comm_temperature_update(self, temp, bedTemp):
+		pass
+
+	def on_comm_state_change(self, state):
+		pass
+
+	def on_comm_message(self, message):
+		pass
+
+	def on_comm_progress(self):
+		pass
+
+	def on_comm_print_job_done(self):
+		pass
+
+	def on_comm_z_change(self, newZ):
+		pass
+
+	def on_comm_file_selected(self, filename, filesize, sd):
+		pass
+
+	def on_comm_sd_state_change(self, sdReady):
+		pass
+
+	def on_comm_sd_files(self, files):
+		pass
+
+	def on_comm_file_transfer_started(self, filename, filesize):
+		pass
+
+	def on_comm_file_transfer_done(self, filename):
+		pass
+
+	def on_comm_force_disconnect(self):
+		pass
